# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OM Pay API",
        "description": "API de paiement mobile OM Pay",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9000/api",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
        "/auth/send-magic-link": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Envoyer un lien magique par SMS",
                "operationId": "d6e016a636c6fb195ccf3ef39037e477",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "phone": {
                                        "description": "Numéro de téléphone au format international",
                                        "type": "string",
                                        "example": "+221777065468"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lien magique généré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lien magique généré"
                                        },
                                        "data": {
                                            "properties": {
                                                "link": {
                                                    "type": "string",
                                                    "example": "myapp://auth/enter-pin?token=abc123"
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "abc123"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Numéro de téléphone non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Téléphone non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/verify-pin": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Vérifier le PIN et obtenir un token d'accès",
                "operationId": "ffcfee17ff694f6158b373f0440fd9ee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "pin"
                                ],
                                "properties": {
                                    "token": {
                                        "description": "Token reçu par SMS",
                                        "type": "string",
                                        "example": "abc123"
                                    },
                                    "pin": {
                                        "description": "PIN à 4 chiffres",
                                        "type": "string",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authentification réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Connexion réussie"
                                        },
                                        "data": {
                                            "properties": {
                                                "access_token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1Qi..."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "PIN invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "PIN invalide"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/send-magic-link-merchant": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Envoyer un lien magique pour la connexion marchant",
                "operationId": "6fce96c7aec2afae4814770880325321",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code_marchant"
                                ],
                                "properties": {
                                    "code_marchant": {
                                        "description": "Code marchand",
                                        "type": "string",
                                        "example": "MCH-000012"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lien magique envoyé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lien magique envoyé"
                                        },
                                        "data": {
                                            "properties": {
                                                "link": {
                                                    "description": "Lien de connexion",
                                                    "type": "string"
                                                },
                                                "token": {
                                                    "description": "Token de session",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Marchant non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Marchant non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/verify-pin-merchant": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Vérifier le PIN pour la connexion marchant",
                "operationId": "3879de937b955c0bb954bb28929f61ae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "pin"
                                ],
                                "properties": {
                                    "token": {
                                        "description": "Token de session",
                                        "type": "string"
                                    },
                                    "pin": {
                                        "description": "PIN à 4 chiffres",
                                        "type": "string",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Connexion réussie"
                                        },
                                        "data": {
                                            "properties": {
                                                "access_token": {
                                                    "description": "Token d'accès JWT",
                                                    "type": "string"
                                                },
                                                "marchant": {
                                                    "$ref": "#/components/schemas/Marchant"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "PIN invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "PIN invalide"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Marchant": {
                "title": "Marchant",
                "description": "Représentation d'un marchant/boutique",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 12
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 45
                    },
                    "code_merchant": {
                        "type": "string",
                        "example": "MCH-000012"
                    },
                    "nom_boutique": {
                        "type": "string",
                        "example": "Boulangerie Diop"
                    },
                    "adresse": {
                        "type": "string",
                        "example": "Rue 12, Dakar Plateau"
                    },
                    "ville": {
                        "type": "string",
                        "example": "Dakar"
                    },
                    "url_qr": {
                        "type": "string",
                        "example": "https://qr.example.com/MCH-000012"
                    },
                    "telephone_service": {
                        "type": "string",
                        "example": "+221781234567"
                    },
                    "email_service": {
                        "type": "string",
                        "example": "contact@boulangeriediop.sn"
                    },
                    "date_activation": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-15T10:30:00Z"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}